Crash fix relating to entity connections
Before this edit, any level that used the function "InteractConnectPropWithMoveObject" would cause a crash when you leave the level, the only exception being when you quit to the main menu without saving. Turned out the save data class for the connections did not have a destructor, meaning the save file processes were calling for an impossible task at times. With this edit, and using integers instead of floats for return functions in scripts when possible, it should hopefully be almost fully playable now.